Move cli authentication to config package
