+++ commands/logs.go
@@ -12,6 +12,7 @@ import (
 	"os"
 	"time"
 
+	"github.com/openfaas/faas-cli/config"
 	"github.com/openfaas/faas-cli/flags"
 	"github.com/openfaas/faas-provider/logs"
 
@@ -96,8 +97,7 @@ func runLogs(cmd *cobra.Command, args []string) error {
 	}
 
 	logRequest := logRequestFromFlags(cmd, args)
-
-	cliAuth := NewCLIAuth(logFlagValues.token, gatewayAddress)
+	cliAuth := config.NewCLIAuth(logFlagValues.token, gatewayAddress)
 	transport := getLogStreamingTransport(tlsInsecure)
 	cliClient := proxy.NewClient(cliAuth, gatewayAddress, transport, nil)
