+++ commands/store_deploy.go
@@ -9,6 +9,7 @@ import (
 	"log"
 	"os"
 
+	"github.com/openfaas/faas-cli/config"
 	"github.com/openfaas/faas-cli/proxy"
 	"github.com/spf13/cobra"
 )
@@ -129,7 +130,7 @@ func runStoreDeploy(cmd *cobra.Command, args []string) error {
 	}
 
 	gateway = getGatewayURL(gateway, defaultGateway, "", os.Getenv(openFaaSURLEnvironment))
-	cliAuth := NewCLIAuth(token, gateway)
+	cliAuth := config.NewCLIAuth(token, gateway)
 	transport := GetDefaultCLITransport(tlsInsecure, &commandTimeout)
 	proxyClient := proxy.NewClient(cliAuth, gateway, transport, &commandTimeout)
